requirements.txt.txt
python-telegram-bot==22.1
requests
schedule
